backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
sentence-transformers==2.2.2
faiss-cpu==1.9.0.post1
pydantic==2.5.0
python-multipart==0.0.6
cryptography==41.0.7
pypdf2==3.0.1
pdfplumber==0.10.3
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.0
transformers==4.35.0
torch==2.6.0
pytest==7.4.3
hypothesis==6.92.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
pyyaml==6.0.1
